easygit: added easygit tool
I know there are tools like gita and stuff which do the same thing
but this tool gets the job done and the output is pretty simple
and easy to read.

When I am working with multiple git repos under my workspace
which can sometimes have multiple branches (obviously) and
multiple remotes as well, its a pain to keep all of them in
track.